Extract default config generation to utils/configuration.ts
This is useful to the qutebrowser port but deserves a live of its own.
